+++ src/connectedvmware/HISTORY.rst
@@ -5,6 +5,7 @@ Release History
 0.2.2
 ++++++
 * GA release: Using stable API version for all operations.
+* `az connectedvmware inventory item show` shows inventory-type specific properties.
 
 0.2.1
 ++++++